Delete: Remoicate shufflenetv2 files
