+++ sscma/models/backbones/__init__.py
@@ -3,8 +3,7 @@ from .EfficientNet import EfficientNet
 from .MobileNetv2 import MobileNetv2
 from .MobileNetv3 import MobileNetV3
 from .pfld_mobilenet_v2 import PfldMobileNetV2
-from .shufflenetv2 import CustomShuffleNetV2, FastShuffleNetV2
-from .ShuffleNetV2 import ShuffleNetV2
+from .ShuffleNetV2 import ShuffleNetV2,CustomShuffleNetV2, FastShuffleNetV2
 from .SoundNet import SoundNetRaw
 from .SqueezeNet import SqueezeNet
 from .MicroNet import MicroNet
